bn_data/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiodns==3.1.1
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
frozenlist==1.4.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyarrow==17.0.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1
